listo a entregar

+++ ft_strlcat.c
@@ -6,7 +6,7 @@
 /*   By: LeoMoreno <lmoreno@student.42quebec.>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/04 17:39:15 by LeoMoreno         #+#    #+#             */
-/*   Updated: 2021/10/04 17:39:59 by LeoMoreno        ###   ########.fr       */
+/*   Updated: 2021/10/14 18:40:59 by LeoMoreno        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,14 +14,14 @@
 
 size_t	ft_strlcat(char *dst, const char *src, size_t dstsize)
 {
-	int	dlen;
-	int	i;
+	size_t	dlen;
+	size_t	i;
 
 	i = 0;
 	dlen = ft_strlen(dst);
-	if ((int) dstsize <= dlen)
-		return ((int) dstsize + ft_strlen(src));
-	while (src[i] != '\0' && ((int) dstsize - 1) > dlen)
+	if (dstsize <= dlen)
+		return (dstsize + ft_strlen(src));
+	while (src[i] != '\0' && (dstsize - 1) > dlen)
 	{
 		dst[dlen] = src[i];
 		dlen++;
